refactor(agent): extract kubernetes module (k3s lifecycle + RPC)
Pull k3s/Kubernetes out of linux/mod.rs into kubernetes.rs:

- 5 RPC handlers: Start / Stop / Delete / Status / Kubeconfig.
- Lifecycle drivers: do_start_kubernetes (data mount + containerd ready
  + spawn k3s server), do_stop_kubernetes(_locked), do_delete_kubernetes
  (clear containerd namespace + wipe k3s/kubelet/cni data dirs).
- Status collector + kubernetes_control_lock serialising mutations.
- pidfile helpers (read/write/remove/process_alive/k3s_pid),
  k3s_api_ready TCP probe.
- K3S_PID_FILE / K3S_NAMESPACE / KUBERNETES_HOST_ENDPOINT constants
  and REQUIRED_KUBERNETES_BINARIES + missing_kubernetes_binaries_at.

Imports the runtime helpers it shares with the Docker path
(daemon_log_file, ensure_containerd_ready, ensure_runtime_prerequisites,
runtime_path_env, missing_binaries_at, CONTAINERD_SOCKET_CANDIDATES).

Side-fix: runtime.rs now reaches `ensure_runtime` via
`crate::agent::ensure_runtime` since linux/mod.rs no longer needs the
re-export now that the runtime + kubernetes modules pull it in directly.

linux/mod.rs drops the heavy paths/ports/protocol use list; only
RequestResult, the Agent struct, handle_connection, and the residual
RPC dispatch remain. Same warning set as before.
