+++ guest/arcbox-agent/src/agent/linux/runtime.rs
@@ -24,9 +24,9 @@ use arcbox_protocol::agent::{
 
 use super::btrfs::ensure_data_mount;
 use super::cmdline::docker_api_vsock_port;
-use super::ensure_runtime;
 use super::probe::{probe_docker_api_ready, probe_first_ready_socket, probe_unix_socket};
 use super::sync_clock_from_host;
+use crate::agent::ensure_runtime;
 use crate::rpc::RpcResponse;
 
 /// Containerd socket candidates (primary + legacy fallback).
